good

+++ main.py
@@ -37,21 +37,41 @@ def job_per_hour(syncer: DataSyncClient):
 
 def _job_for_cinema_tickets_hourly(syncer: DataSyncClient):
     today = date.today().strftime("%Y-%m-%d")
-    dt_5am = datetime.combine(date.today(), time(6, 0, 0))
-    table_name = syncer.config.get_name('C01')
+    current_hour = datetime.now().hour
 
-    list_of_ids = syncer.lark_client.get_table_records_id_after_date(table_name, dt_5am, syncer._get_primary_timestamp_column_name('C01'))
-    syncer.lark_client.delete_records_by_id(table_name, list_of_ids)
-    query_data_today = FinancialQueries('C01', 'day', today)
-    syncer.upload_data(query_data_today, table_name)
+    if current_hour > 6:
+        dt_5am = datetime.combine(date.today(), time(6, 0, 0))
+        table_name = syncer.config.get_name('C01')
+
+        list_of_ids = syncer.lark_client.get_table_records_id_after_date(table_name, dt_5am, syncer._get_primary_timestamp_column_name('C01'))
+        syncer.lark_client.delete_records_by_id(table_name, list_of_ids)
+
+        query_data_today = FinancialQueries('C01', 'day', today)
+        syncer.upload_data(query_data_today, table_name)
+    
+    if current_hour < 6:
+        yesterday = (date.today() - timedelta(days=1)).strftime("%Y-%m-%d")
+        dt_5am = datetime.combine(date.today() - timedelta(days=1), time(6, 0, 0))
+        table_name = syncer.config.get_name('C01')
+
+        list_of_ids = syncer.lark_client.get_table_records_id_after_date(table_name, dt_5am, syncer._get_primary_timestamp_column_name('C01'))
+        syncer.lark_client.delete_records_by_id(table_name, list_of_ids)
+
+        query_data_today = FinancialQueries('C01', 'day', yesterday)
+        syncer.upload_data(query_data_today, table_name)
+
+
     _message_after_tickets_job(syncer)
 
 def _job_for_cinema_ticket_daily(syncer: DataSyncClient):
+    today = date.today().strftime("%Y-%m-%d")
     dt_5am = datetime.combine(date.today() - timedelta(days=14), time(6, 0, 0))
 
     table_name = syncer.config.get_name('C01')
     list_of_ids = syncer.lark_client.get_table_records_id_before_date(table_name, dt_5am, syncer._get_primary_timestamp_column_name('C01'))
     syncer.lark_client.delete_records_by_id(table_name, list_of_ids)
+    query_data_today = FinancialQueries('C01', 'day', today)
+    syncer.upload_data(query_data_today, table_name)
 
 def _job_for_others(syncer: DataSyncClient):
     syncer.sync_all_yesterday()
@@ -84,11 +104,6 @@ def _message_after_job(syncer: DataSyncClient):
         syncer.lark_client.get_chat_group_id_by_name('服务器状态')
     )
 
-    syncer.lark_client.send_message_to_chat_group(
-        message,
-        syncer.lark_client.get_chat_group_id_by_name('服务器状态')
-    )
-
 def _message_init(syncer: DataSyncClient):
     current_time = datetime.now()
     timestamp = current_time.strftime("%Y-%m-%d %H:%M:%S")[:-3]
@@ -114,7 +129,7 @@ if __name__ == "__main__":
         scheduler.add_job(
             job_per_hour,
             'cron',
-            hour='0,8-23',
+            hour='8-23,0-5',
             minute=0,
             args=[global_syncer]
         )
